+++ projects/_03_independent_cqc/_04_model/tests/utils/test_versioning.py
@@ -1,7 +1,9 @@
+import io
 import json
 import unittest
 
 import boto3
+import joblib
 from moto import mock_aws
 
 from projects._03_independent_cqc._04_model.utils import versioning as job
@@ -48,19 +50,24 @@ class GetRunNumberTests(unittest.TestCase):
         self.assertEqual(run_number, 3)
 
 
-class SaveMetadataTests(unittest.TestCase):
+class SaveModelAndMetadataTests(unittest.TestCase):
+    def setUp(self) -> None:
+        self.s3_root = "s3://pipeline-resources/models/model_A/"
+        self.test_model = {"coef": [1.0, 2.0, 3.0]}
+
     @mock_aws
-    def test_save_metadata_creates_metadata_file(self):
+    def test_creates_metadata_file(self):
         s3 = boto3.client("s3", region_name="eu-west-2")
         s3.create_bucket(
             Bucket="pipeline-resources",
             CreateBucketConfiguration={"LocationConstraint": "eu-west-2"},
         )
 
-        s3_root = "s3://pipeline-resources/models/model_A/"
-
         metadata = {"metric": 0.99}
-        job.save_metadata(s3_root, run_number=5, metadata=metadata)
+
+        job.save_model_and_metadata(
+            self.s3_root, run_number=5, model=self.test_model, metadata=metadata
+        )
 
         # Read back the object
         response = s3.get_object(
@@ -70,3 +77,59 @@ class SaveMetadataTests(unittest.TestCase):
 
         self.assertIn("timestamp", body)
         self.assertEqual(body["metric"], 0.99)
+        self.assertEqual(body["run_number"], 5)
+
+    @mock_aws
+    def test_creates_model_file(self):
+        s3 = boto3.client("s3", region_name="eu-west-2")
+        s3.create_bucket(
+            Bucket="pipeline-resources",
+            CreateBucketConfiguration={"LocationConstraint": "eu-west-2"},
+        )
+
+        job.save_model_and_metadata(
+            s3_root=self.s3_root,
+            run_number=7,
+            model=self.test_model,
+            metadata={},
+        )
+
+        response = s3.get_object(
+            Bucket="pipeline-resources",
+            Key="models/model_A/7/model.pkl",
+        )
+
+        model_bytes = response["Body"].read()
+        loaded_model = joblib.load(io.BytesIO(model_bytes))
+
+        self.assertEqual(loaded_model, self.test_model)
+
+    @mock_aws
+    def test_objects_are_saved_under_correct_run_prefix(self):
+        s3 = boto3.client("s3", region_name="eu-west-2")
+        s3.create_bucket(
+            Bucket="pipeline-resources",
+            CreateBucketConfiguration={"LocationConstraint": "eu-west-2"},
+        )
+
+        job.save_model_and_metadata(
+            s3_root=self.s3_root,
+            run_number=42,
+            model=self.test_model,
+            metadata={},
+        )
+
+        objects = s3.list_objects_v2(
+            Bucket="pipeline-resources",
+            Prefix="models/model_A/42/",
+        )
+
+        keys = {obj["Key"] for obj in objects.get("Contents", [])}
+
+        self.assertEqual(
+            keys,
+            {
+                "models/model_A/42/model.pkl",
+                "models/model_A/42/metadata.json",
+            },
+        )