+++ x_non_pipeline_utils/add_sector_and_service_to_CQC_register.py
@@ -1,28 +1,31 @@
+from pathlib import Path
+
 import pandas as pd
 from cqc_metadata import ColumnNames as Columns
 from cqc_metadata import ColumnValues as Values
 from cqc_metadata import CqcCategories, CqcConfig
 
+YEAR_AND_FILE_NAME = Path("2025/12. CQC 051225 (from CQC website)")
+
 
 def main():
-    file = CqcConfig.directory / CqcConfig.old_file_name
-    data = open_CQC_File(file, CqcConfig.sheet_name)
+    file = Path(f"{CqcConfig.directory / YEAR_AND_FILE_NAME}{CqcConfig.source_suffix}")
+    data = open_cqc_file(file, CqcConfig.sheet_name)
     data = remove_non_social_care_data(data)
     data = add_sector_data(data)
     data = add_service_data(data)
-    data = save_CQC_file(data)
+    data = save_cqc_file(data)
     print("complete")
 
 
-def open_CQC_File(File_name, sheet_name):
-    print("opening file")
-    CQCdata = pd.read_excel(
-        File_name, sheet_name=sheet_name, skiprows=CqcConfig.blank_rows
-    )
+def open_cqc_file(file_name, sheet_name):
+    print(f"opening file: {file_name}")
+    CQCdata = pd.read_excel(file_name, sheet_name=sheet_name)
     return CQCdata
 
 
 def remove_non_social_care_data(data):
+    print("removing non social care data")
     data = data[data[Columns.location_type] == Values.social_care_org].reset_index(
         drop=True
     )
@@ -68,9 +71,11 @@ def add_service_data(data):
     return data
 
 
-def save_CQC_file(df):
-    print("saving file")
-    save_location = CqcConfig.directory / CqcConfig.new_file_name
+def save_cqc_file(df):
+    save_location = Path(
+        f"{CqcConfig.directory / YEAR_AND_FILE_NAME}{CqcConfig.dest_suffix}"
+    )
+    print(f"saving file: {save_location}")
     df.to_excel(save_location, sheet_name=CqcConfig.new_sheet_name, index=False)