+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Called the forward_fill_latest_known_value function on pir_people_directly_employed_dedup column to replicate the forward fill process for PIR Data.
 
+- Polars job [model_02_train](projects/_03_independent_cqc/_04_model/fargate/model_02_train.py) to train, test and save a specified model.
+
 - A new function to remove longitudinal outliers from CT data. The function flags the outliers based on absolute difference in median values and removes the ourlier value. 
 
 ### Changed
@@ -15,6 +17,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 -
+
 
 ## [v2025.11.0] - 18/12/2025